requirements.txt
tensorflow>=2.12.0
librosa>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
soundfile>=0.12.0
tqdm>=4.62.0
PyYAML>=6.0
unidecode>=1.3.0
transformers>=4.21.0
wandb>=0.15.0
psutil>=5.9.0
pynvml>=11.4.1

# New dependencies for evaluation and optimization
tensorflow-model-optimization>=0.7.0
matplotlib>=3.5.0

# Benchmarking and monitoring dependencies
GPUtil>=1.4.0
seaborn>=0.11.0
plotly>=5.0.0